Define a plugin API for CxxLibraryDescription
Summary:
This introduces a plugin API for `CxxLibraryDescription` so that we can change the implementation of `AppleLibraryDescription` to not use a Swift delegate pattern. The reason is that the existing approach with a Swift delegate pattern exhibits significant drawbacks, documented in [issue #1472](https://github.com/facebook/buck/issues/1472).

After further prototyping and discussion with robbertvanginkel, it became clear that we can achieve most of our goals without a major rewrite as suggested previously.

In particular, the API allows the injection of:

- `CxxPreprocessorInput`: This will be used to expose `Module.swiftmodule` and Obj-C generated header via a framework-style import (e.g., `<Module/Module-Swift.h>`).
- Object Files: Object files from the Swift compilation process.
- Linker Arguments: Linker arguments necessary to link any final products which transitively depend on mixed Obj-C + Swift `apple_library`.

Test Plan: Existing unit tests as there's no change in behavior as there's no plugin instance.

Reviewed By: benb

fbshipit-source-id: 0b2ace7