Don't use test_host as bundle_loader for UI tests
Summary:
This allows `apple_test` bundles with `is_ui_test` to be correctly build
with buck, so that they can be used with an external test runner.

Currently, the `test_host_app` is used as the `bundle_loader` when
linking an apple_test. Since the uitest bundle isn't injected into the
app it is testing, but into a stub app when building a uitest we should
not use it as the bundle_loader.
Closes https://github.com/facebook/buck/pull/1505

Reviewed By: yiding

Pulled By: yiding

fbshipit-source-id: c972f47