Suggest next steps for compiler crashes
Summary:
I've run into a few cases where `javac` crashes when run entirely without deps, in a way
that Buck itself cannot completely prevent. For the most part these cases represent an
error in the code (e.g. a static import of a type that does not exist even with deps), so
this commit suggests developers should build the real target to find the problem.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: a2289a3