PCH: CxxPrecompiledHeaderTemplate rules should not force shared/static linkage
Summary:
This was previously setting linkage to `SHARED`, but this resulted in builds using PCH rules to be forced to shared, even if they preferred static themselves.

There's an `ANY` enum value available, this rule type should be using that.  (In `AbstractPrebuiltCxxLibraryGroupDescription` the logic is such that header-only libs use `ANY`, and this should follow suit.)

Test Plan:
* Local tests w/ `ant -Dtest.class=\*CxxPrecom\*Test java-test`
* Watching our CI

Reviewed By: andrewjcg

fbshipit-source-id: 3f4404d