Remove ProjectFilesystem dependency on util.zip
Summary:
`ProjectFilesystem` is a low level abstraction that provides access to a filesystem. At the same time it has a method to create a zip file. Zip functionality is on the next logical level because it operates on files and directories. For example, Unzip class uses `Projectfilesystem` and since `ProjectFilesystem` depends on `util.zip` it's not possible to move `Unzip` to `util.zip`.

This diff moves zip-related functionality from `ProjectFilesystem` to `util.zip.Zip` class. Another change in this commit is renaming `testutil.Zip` to `testutil.ZipArchive` to avoid ambiguity.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 4f87c31