Make error suppression more targeted
Summary:
Turns out there are some errors that can occur during Enter from which the compiler kinda recovers
but with incomplete or otherwise corrupted data structures. (For example, the compiler does fine if a class's
superclass is missing, but if the superclass is present but *its* superclass is missing it will
only partially complete the type.) This commit changes our error suppression strategy to only
suppress the "recoverable" errors -- that is, the ones that the compiler itself defers until after annotation
processing.

Test Plan: Updated unit tests

Reviewed By: asp2insp

fbshipit-source-id: 96172e2