Add some tests for constants defined in terms of other constants that are missing
Summary:
Before the suppression strategy change in my previous commit, the source-ABI generation
was running outside of error suppression. When it called `VariableElement.getConstantValue` on
a constant defined in terms of a missing constant, that would cause a "recoverable" compiler
error to be emitted, and the tests added by this commit would fail. With the new suppression
strategy, things work right.

Test Plan: These are new tests

Reviewed By: dreiss

fbshipit-source-id: f4d9acc