Add a regex symbol whitelist
Summary:
It is currently impossible to make the relinker preserve specific symbols.
This is mostly not a problem since the relinker is always right for strong
dependencies. When using dynamic (`dlsym(3)`) or weak symbols, however, this
breaks down.

We have multiple cases that require a relinker whitelist now, so this diff adds
one, as a list of regexes over the symbol names.

fbshipit-source-id: b92b1a1