Use DefaultProjectFilesystem in CachingBuildEngineTest
Summary:
FakeProjectFilesystem differs in ways that make it difficult to
correctly test the CachingBuildEngine. Change these tests to use a
DefaultProjectFilesystem instead.

Updates all the tests to create directories for files that they try to
write.

Makes one test close the CachingBuildEngine before shutting down the
service (when using a real filesystem, some of the async tasks get
interrupted as expected but the test doesn't handle that correctly).

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: d895546