Cache InitializeFromDisk state between builds
Summary:
This adds an invalidate() method to BuildOutputInitializer. This is
called by CachingBuildRuleBuilder before changing the outputs of a
buildrule. Then, the initializer will use the cached state if it's there
(so, the first time a rule is built in a daemon it will be initialized
and then only need to be re-initialized if it is fetched from cache or
built locally).

Test Plan:
CI. Added tests for all BuildRuleSuccessType that verify a rule is
initialized from disk as expected and that it is invalidated for the
cached/local success types.

Reviewed By: ttsugriy

fbshipit-source-id: 1b2c665