Give NonPredexedDexBuildable its own output dirs
Summary:
This makes NonPredexedDexBuildable write its outputs into its own
directories not referenced directly by others.

Adds the proguard text files directory to DexFilesInfo and then
AndroidBinaryBuildable copies that off into a separate directory for it
to refer to.

This is in preparation for splitting this out into its BuildRule.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 85ac2e5