Parallelize parts of CxxLinkableEnhancer.createCxxLinkableBuildRule
Summary:
Create an interface on BuildRuleResolver for conditionally converting a
stream to parallel stream. This will be used as the primary method to
introduce concurrency to large tasks during action graph creation.

This is better than exposing ExecutorService-like APIs, as ForkJoinPool
is somewhat sensitive to the topology of the submitted tasks. Parallel
streams takes this into account and generate tasks internally that works
better with ForkJoinPool.

Use this interface to parallelize
CxxLinkableEnhancer.createCxxLinkableBuildRule, which will allow for the
creation of cxx libraries (static and shared) to be done in parallel.

Test Plan:
run buck targets --show-rulekeys on c++ binaries, see that the
rulekeys are the same as before, both with action graph parallelizing on and
off.

Reviewed By: ttsugriy

fbshipit-source-id: aa884ed