Make ModernBuildRule support self-Buildables
Summary:
For simple BuildRules, there may be no fields in the BuildRule that need
to be excluded from the Buildable. In that case, forcing the Buildable
implementation to be a different class is unnecessary.

This diff provides a way to construct a ModernBuildRule such that the
BuildRule implementation is also the Buildable implementation.

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: a99270d