Move NdkCxxPlatformsProvider creation logic to NdkCxxPlatformsProviderFactory
Summary:
`android/toolchain:toolchain` has a lot of dependencies which are there because we have implementation of `NdkCxxPlatformsProvider` in it. This blocks further refactoring because it introduces circular dependencies.

To reduce the amount of dependencies in this target the implementation of `NdkCxxPlatformsProvider.create` was moved to `NdkCxxPlatformsProviderFactory` in `android/toolchain/impl:impl`.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 4361418