Give coordinator time to return final minion response before shutting down Thrift server
Summary:
When the build is complete, on the final minion request for more work the coordinator will set a Future that initiates its own shutdown process. This generally completes before the response has been written back to the minion's socket, and as a result the minion crashes with a Thrift request failed exception.

To mitigate this problem, introducing a 100 millisecond sleep before shutting down the Thrift server.

Test Plan: ci

Reviewed By: shivanker

fbshipit-source-id: 2bec739