addToIndex completes the pending task if one already exists
Summary:
Instead of being a no-op, addToIndex actually completes the pending task if one
exists.

This is used to support the case where a rule's construction is deferred to
another rule, which constructs the first rule and sets it via addToIndex, and
then the first rule returns the value via getRule. This pattern is used in
android_binary.

Test Plan: buck targets --show-output android_binary android_binary#unstripped_native_libraries

fbshipit-source-id: 1dc9371