Fix creation of build target for Swift compile rules
Summary: The build targets passed to create `SwiftCompiler` rules can include additional flavors that are not part of `apple_library` (e.g., `AppleLibraryDescription` is composed by `AppleTestDescription`). This means that in order to create the correct flavor for the `SwiftCompile` rules, we need to just include the platform and the compile flavor and exclude everything else.

Test Plan: Existing integration tests

Reviewed By: dinhviethoa

fbshipit-source-id: 989ed6e