Provide API to inject CxxPreprocessorInput
Summary: When compiling Swift code as part of `AppleLibraryDescription`, it assumes it can get a placeholder node (`CxxLibrary`) to be able to get preprocessor inputs. This assumption does not hold true for `AppleTestDescription`. In this diff, we provide an API (unused here) so that this process can be customised.

Test Plan: Existing integration test

Reviewed By: dinhviethoa

fbshipit-source-id: 536989e