Move low level Config-related classes from config:config to util/config:config
Summary:
This a preparation to move `cli.BuckConfig` to `config` package.

`config` package contains basic `Config` and related classes, but `BuckConfig` is more abstract class and should be separated from lower level classes.

In this change most of classes in `config` package are moved to `util.config`. These classes are mostly internal and not widely used as `BuckConfig`.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 0e35e72