Remove step:step dependency on event:event
Summary: `event:event` has lots of dependencies which are relevant to the implementation, but not to the interface. At the same time `ConsoleEvent` and `ThrowableConsoleEvent` are implementations of `BuckEvent`, but they don't have implementation dependencies so they can live in `interface` target. This allows us remove the `step:step` dependency on `event:event` and use `event:interfaces` instead.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 6f82301