Hide creation of ProjectFilesystem inside TestProjectFilesystemHelper in tests
Summary:
This is the preparation to move `ProjectFilesystem` creation logic outside of the class itself. See the following commit for more information.

In this change I'm adding a test helper class that hides the creation of `ProjectFilesystem` instances. This will let me replace this logic in the following diff without touching these test classes.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 99e9984