Move creation of ProjectFilesystem to ProjectFilesystemFactory
Summary:
`ProjectFilesystem` is the basic low-level object which is used almost everywhere, but at the same time it has lots of dependencies which makes it difficult to refactor components that depend on it. Most of these dependencies are required for implementation, but not for the interface.

In this change the logic to create `ProjectFilesystem` is moved outside to `ProjectFilesystemFactory`. It allows us remove some dependencies in the following changes.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: e2b2a5e