JavaBinary/AndroidBinary: Add tests for behaviors around multiply-specified deps
Summary:
It is possible for a dependency of a `java_library`/`android_library` to appear in more than one
of its dependency lists (`deps`, `exported_deps`, and/or `provided_deps`). When this occurs,
`java_binary` and `android_binary` handle the situation differently.

`java_binary` includes anything that is present in `deps` or `exported_deps`, while `android_binary`
includes anything that is present in `deps` or `exported_deps` but *NOT* `provided_deps`. Effectively
`java_binary` considers there to be an order of precedent among the lists, whereas `android_binary`
considers them as modifying one another.

This behavior is confusing and should be changed, but as I'm changing how things are constructed
I need to preserve it, so I'm adding some tests for it.

Test Plan: These are new tests

Reviewed By: dreiss

fbshipit-source-id: e711504