Decouple buck binary built with ant from repo
Summary:
If buck binary is built with ant, switching to a different
revision can cause it to use jars from a different commit which
will can break. This change also autogenerates classpaths file.
In follow-up diffs it would be great to get rid of manual classpaths file
in `programs` directory.

Test Plan: CI

Reviewed By: styurin

fbshipit-source-id: 3b19e95