Change Build.executeAndPrintBuildFailures to print failures for all exceptions
Summary:
Previously, some exceptions from the build were handled here and
others were just propagated out of this function and reported to the user by
something else (typically by nailgun as it propagates all the way out of main).
So that all errors from the build are reported to the user in a consistent way,
do all the logging here.

Test Plan:
BuildCommandErrorsIntegrationTest is updated and shows how this
changes output of various scenarios.

Reviewed By: ttsugriy

fbshipit-source-id: ae072ee