Simplify Build.rootCauseOfBuildException
Summary:
This function had some odd behavior where it would sometimes return an
exception that would then be thrown and sometimes the function itself would
throw an exception. It was also wrapping some functions in RuntimeException
(adding some rather unnecessary stack frames to reports). This changes the
function to just return the root cause and then let that be thrown by the caller.

Test Plan:
BuildCommandErrorsIntegrationTest is updated and shows how this
changes output of various scenarios.

Reviewed By: ttsugriy

fbshipit-source-id: 2a4ca5f