Add ErrorLogger and related interfaces
Summary:
This adds an ErrorLogger that can handle extracting a good user
message out of an exception chain. It adds two interfaces, WrapsException and
ExceptionWithContext, to assist in that process.

Basically, when logging an exception, ErrorLogger will unwrap through all the
outermost WrapsException exceptions accumulating context from those that have
context and then construct log to the console a message from the underlying
exception. If the exception is a HumanReadableException, it will be logged as a
user error and if it's any other exception type it will be logged as an
internal error. In a following diff, you can see how some of that logic can be
customized.

Also adds BuckExecutionException and BuckUncheckedExecutionException as good
ways of wrapping exceptions that can't be thrown from a function. These should
be preferred to doing `new RuntimeException(exception)`.

Test Plan: Added simple test.

Reviewed By: ttsugriy

fbshipit-source-id: 4cf9090