Use ErrorLogger in Build.executeBuildAndPrintFailures
Summary:
This uses the new ErrorLogger in Build which simplifies the error
output and allows the caching build engine to attach context to exceptions in a
good way.

The ErrorLogger behavior is slightly customized to match the previous behavior
of this function.

IOException and StepFailedExceptions are handled specially so
that they are reported as user errors instead of internal errors.

For internal errors, we include the full stack trace in the message instead of
just the message.

Test Plan:
BuildCommandErrorsIntegrationTest is updated and shows how this
changes output of various scenarios.

Reviewed By: ttsugriy

fbshipit-source-id: f81debd