Use correct platform for actool when we're building a watchos asset catalog
Summary:
When building a watchos legacy app, all the flavors from the main app are propagated down. As a result, the assets catalog are passed a platform that doesn't match watchos because it picks one of the platforms.
This diff forces the correct platform in the case of legacy watchos app.

Reviewed By: ttsugriy, rnystrom

fbshipit-source-id: bac7fef