Change DexFilesInfo to support reading paths from file
Summary:
Once NonPreDexedDexBuildable is in its own BuildRule, it will not be
able to compute the secondary dex paths before the AndroidBinary is
created and so it needs to pass that information in some output file.
This changes DexFilesInfo to support finding the secondary dex dirs in
this way and updates NonPreDexedDexBuildable to pass them that way.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 5b49812