Remove the mistaken test that prevented SWEs from creating Project Views in the repo
Summary: The 'not under the repo' checking, wasn't really right, and couldn't be made right in a portable way. For example, a company might have a master repo with project or architecture specific subsidiary repos in subdirectories. If the SWE is in `master_repo/sub_repo`, the `Cell` path will be  `master_repo/sub_repo`, so that we can detect views under  `master_repo/sub_repo` ... but we can't detect views under `sub_repo` which are just as undesirable.

Reviewed By: mkillianey

fbshipit-source-id: cddc4a0