Keep libraries used in tests in one place to share them between tests
Summary:
We have lots of duplicate libraries in tests.

We can keep them in one place and add a cell with these libraries when needed to access these libraries. In this change `TimeoutIntegrationTest` is migrated to use shared libraries.

Test Plan: CI and local testing with `buck test` and IntelliJ.

Reviewed By: ttsugriy

fbshipit-source-id: 10ed8bd