Added a config to generate unique static library name
Summary:
When projects have lots of dependencies, the dependencies might have similar target names.
In Xcode, at the linking stage, it will result in collision of library names since the names of the library are based on the name of the target.

This patch introduces a config to tell buck to generate libraries with names that are less likely to collide.

Test Plan: unit tests + CI

Reviewed By: bertmaher

fbshipit-source-id: fa48ce2