Create symlinks of the build outputs to buck-out/last
Summary:
It's usually hard to find the location of the artifacts:
we generally need to use --show-output or --show-full-output with buck build.
This diff will symlinks all the build outputs to buck-out/last.
It will make it easier for the user to find the artifacts.

Test Plan: CI and manual test with buck build.

Reviewed By: sbalabanov

fbshipit-source-id: 1c73d23