versions: add better error message for bad version deps
Summary:
Currently, when a description that doesn't support versioning depends
(transitively) on a versioned node, we fail a precondition check and present
the user with an ugly Java stack trace.  This diff makes this error message
a more useful readable exception which includes a proper dependency trace.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 60fcb6a