Remove a ConcurrentHashMap in DefaultRuleKeyCache
Summary:
Instead of storing the dependent information and the value in different
ConcurrentHashMaps, use a single ConcurrentHashMap to store a structure of both
of these fields.

This removes one ConcurrentHashMap at the cost of a new wrapper object, but its
a win in terms of memory used. (Due in part to having fewer IdentityWrapper
objects as well.)

Test Plan: heap dump after a `buck targets`

Reviewed By: dinhviethoa

fbshipit-source-id: 82f5403