Use a memoizing supplier that drops reference to the delegate supplier in DefaultRuleKeyCache
Summary:
Introduce "MoreSuppliers.memoize", which drops reference to the original
supplier after computing the value. Guava's Suppliers.memoize holds on to the
reference so that the instance can be serialized using java serialization. We
don't usually care about that.

Use these kind of suppliers in DefaultRuleKeyCache, drops the memory used for
this Supplier field from ~70 to ~24 bytes per cache entry.

Test Plan:
Insert a sleep in `TargetsCommand` after computing the rule key but before
RuleKeyFactory is collected., heap dump during this time.

Reviewed By: dinhviethoa

fbshipit-source-id: f715109