Enhance AddsToRuleKey
Summary:
This allows annotating an Immutable with AddsToRuleKey. This
Immutable can then annotate methods with AddToRuleKey. When
constructing the rulekey, the annotated methods will be called and the return
value will be added to the rulekey.

AddToRuleKey can only be used on methods within an Immutable.

This primarily meant as a method to move Immutables off of the
deprecated RuleKeyAppendable.

Test Plan: CI

Reviewed By: jkeljo

fbshipit-source-id: 10120ed