Fix DefaultRuleKeyCache NPE during invalidation
Summary:
Fix regression I introduced where I did not institute a null check when
changing the DefaultRuleKeyCache data structures.

When invalidating nodes in a dag, one path may have invalidated a node that's
queued to be invalidated via another path. Not checking for existence of cache
entry will cause a NPE in this case.

Also, infer eradicate didin't trigger warning for ConcurrentHashMaps, this is
kind of a big deal.

Test Plan: added a test.

Reviewed By: dinhviethoa

fbshipit-source-id: 610e53a