pch: static-linked binaries w/ PCH shouldn't get duplicate deps
Summary:
If a binary / one of its lib deps have a precompiled header and the PCH has a dep which is whole-archive-linked, then that dep's archive may be specified twice in the linker command line.

This would result in a link error:

    stderr: .../tools/binutils/bin/gold/ld: error: .../buck-out/experimental/steveo/pch/lib#static/liblib.a(.../buck-out/experimental/steveo/pch/lib#static/../lib#compile-lib.cpp.ocb207ca1/lib.cpp.o): multiple definition of 'foo()'
    .../tools/binutils/bin/gold/ld: .../buck-out/experimental/steveo/pch/lib#static/liblib.a(.../buck-out/experimental/steveo/pch/lib#static/../lib#compile-lib.cpp.ocb207ca1/lib.cpp.o): previous definition here
    collect2: error: ld returned 1 exit status

It's not two different libraries providing the same symbol, it's the same library just listed twice.

Change so that `getNativeLinkableInput` returns empty (and **not** returning deps there), and letting the deps get returned in other methods `getNativeLinkableDeps`, `getNativeLinkableExportedDeps`.

Test Plan:
* local test added, covering this case (and now works).  `ant -Dtest.class=CxxPrecompiledHeaderRuleTest java-test`
* watching CI

Reviewed By: yiding

fbshipit-source-id: 711586a