Always use Swift delegate in apple_binary
Summary: Since the work on removal of the Swift delegate in `apple_binary` is currently postponed but we still want to switchover to the new behaviour for `apple_library` and `apple_test`, we have to remove the check here and always use the Swift delegate. Otherwise, this would regress existing Swift support in `apple_binary` which can break existing apps.

Test Plan: Existing integration tests

Reviewed By: dinhviethoa

fbshipit-source-id: 37f0cba