(IjProject) Let ignored paths to be added to excluded folders in the project
Summary:
IntelliJ project generation has a part that creates a list of folders excluded folders. This logic is based on a walking a subtree in `ProjectFilesystem`. The implementation of walking always skips ignored files and, as a result, ignored files aren't added to excluded folders in IntelliJ projects.

The fix is to allow walking through ignored paths.

Test Plan: Build a project, check `project_root.iml`, it should exclude `buck-out/.trash`.

Reviewed By: JonShemitz

fbshipit-source-id: d30abde