Fix excessive concurrency when running tests with buck internal test runner
Summary:
I inadvertently increased concurrency of tests run with buck internal test runner.

It used to be limited by virtual cpu resoures (num processors), but my changes
caused it to be no longer resource aware, and it was running with the
concurrency equal to number of threads in the pool instead. This is usually
higher than num processors. This contributes to test flakiness.

A more comprehensive fix will come in the future.

Test Plan: visualscrutiny

fbshipit-source-id: cea07c4