experimental support for raw_headers
Summary:
A new attribute raw_headers is added to cxx_library and other cxx rules. Technically, buck uses raw_headers for two purposes:
To associate a header with cxx library(-ies)
1. To use it during header verification.
2. Raw headers are used as "they are": they are not copied or symlinked by buck. Buck doesn't add additional -I flags for such headers.

`raw_headers` are considered to be experimental for now and cannot be used together with traditional headers in the same rule.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: f0d7e14