`JsBundleGenrule`: Fix implementation of `AndroidPackageable`
Summary:
`JsBundleGenrule` did not only forward the resources exposed by the underlying bundle, but also its JavaScript output. That means that the output of the genrule itself was always discarded.

This change makes sure that the output of a `js_bundle_genrule` ends up in `android_binary` outputs as expected.

Test Plan:
- new integration test
- fixed expectations in unit test

Reviewed By: cpojer

fbshipit-source-id: 6ec0c6c