Make -ObjC linker flag configurable
Summary: The `-ObjC` flag can be undesirable in certain cases. There is no linker flag that negates `-ObjC`, so there needs to be a way to prevent buck from adding `-ObjC`. This change adds a config option, `apple.link_all_objc`, that can be used to disable the `-ObjC` flag.

Test Plan: Sandcastle

Reviewed By: gunnarku, dinhviethoa

fbshipit-source-id: e5698c6