Use URI of jar file instead of encoded URL in JavaInMemoryFileManager
Summary:
Apparently using `URL.encode` works fine on Linux and OS X, but not on Windows. This forces `ClassUsageTracker` to throw an exception when an annotation processor is trying to access a resource because the result URI is malformed.

The fix is to use Path's URI  which should be encoded properly. Also, changed a logic to save the common prefix to reduce the amount of string concatenations.

Test Plan: Changes down the stack pass.

Reviewed By: ttsugriy

fbshipit-source-id: 0a4538b