TraceAnalyzer: Buffer events in a queue
Summary:
Events are not guaranteed to be in timestamp order, but we want to
process them that way.  Buffering a whole trace into memory and sorting
would take a scary amount of memory.  Instead, buffer a fixed number of
events into a priority queue and keep them sorted as a sliding window.
If the window isn't big enough, we fail.

Test Plan: Ran it.

Reviewed By: bertmaher

fbshipit-source-id: c0485f3