Add rule to compute exo-resources files hashes
Summary:
These hashes are used for both the metadata.txt created for the
installed files and for determining the paths to the files themselves. Right
now, the files are hashed on every install invocation. This adds a rule that
just hashes an input file and writes the hash to the output path. Then,
exopackage installation reads those hashes instead of computing them itself.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 70f7f4f