Deleting the binary in link rule should be recursive
Summary: When switching target types with the same name, we might end up creating something in the desired binary's location that's actually a folder. Need to recursively delete that just in case.

Test Plan: Build a target as a ghci-rule to success, switch target to a binary-rule, and watch it successfully rebuild.

Reviewed By: niteria

fbshipit-source-id: 64adb79