pch: offer failover mode if precompile unsupported or disabled
Summary:
For some builds it may be necessary to turn off PCH, due to some alternate toolchain not supporting PCH.  For those cases, there would be an error if the target graph still tries to use PCHs.  One might want to just change the error to a warning and proceed anyway.

This diff moves some of the PCH logic from `requirePreprocessAndBuildRule`; this logic was a bit convoluted, and now gets a little more complex, so it's now cleaned up a little and lives in a different method.

Test Plan:
* local build + test: `ant  &&  ant -Dtest.class=Cxx*Pre*Test java-test` (New test cases added)
* `NO_BUCKD=1 ./bin/buck build buck`
* watching CI

Reviewed By: dinhviethoa

fbshipit-source-id: bbe7766