Add --rulekeys-log-path to buck build and buck targets
Summary: This adds a --log-rulekeys option to buck build and buck targets. This enables the thrift hasher in places where we are hashing a rule key and writes it out in a binary format to the provided file. We can then build tooling on the top of this to extract and examine the action graph in a structured way.

Test Plan:
Ran buck build and buck targets with the given flags, and I was able to use a sample tool to extract the data back out after the run.
Added some integration tests to make sure that the binary data gets written out.

Reviewed By: ttsugriy

fbshipit-source-id: 5c6c9a3