Refactor TargetsCommand to put result accumulator into a small class
Summary:
I was investigating something and chanced upon this file and found some oddities.

"ShowOptions" is actually an computed fields that the user requested, and we
use ShowOptions.Builder to accumulate the values of these different requested
fields. The map that was passed around is actually used to accumulate these
objects for all requested targets. I've renamed the classes and wrapped the Map
in a class, capturing the common access pattern.

Also renamed the enum field enumeration and attached the getter function for
each field, so filling the json object need only loop through the enum fields,
instead of copy pasta.

Moved creating the filtered attribute map for each target down to where the map
is passed to the json serializer. This way we don't have to explicitly check whether
a field is filtered out before inserting, which would have caused runtime exceptions
previously, if we inserted such a field into the filtered view of the map.

Test Plan: CI

Reviewed By: dinhviethoa

fbshipit-source-id: 1099cf8