Fix regression where build info was no longer passed in structured way to Chrome trace.
Summary:
This reasserts the logic that was added in D5568069 (and was removed in
D5955726) that made it possible to reliably extract certain data from the Chrome
trace using a JSON parser.

This also reintroduces the test case to verify this behavior that was removed in
D5568069.

Test Plan: `buck test test/com/facebook/buck/event/listener`

Reviewed By: ttsugriy

fbshipit-source-id: 8143128