read version on both stderr and stdout for real
Summary:
It turns out the logic was wrong and we would not read stderr when stdout was
empty. Infer still outputs the version info on stderr.

Test Plan: Changed the test to output to stderr, checked that it works for both with `buck test //test/com/facebook/buck/cxx:cxx-binary-integration`

Reviewed By: martinoluca

fbshipit-source-id: b8a4018