Specify executors in Futures.transform for expensive tasks.
Summary: There were a lot of places in the source file materialization code-path, where we were using Futures.transform and Futures.transformAsync without specifying an executor -- which is not suggested for anything other than super-light-weight callbacks. This diff makes sure we pass the global CPU thread pool for the same.

Test Plan: CI

Reviewed By: aiked

fbshipit-source-id: e5dc5fe