Jsr199JavacInvocation: Switch back to `JavacTask.call`
Summary:
Calling the phases individually was a nice idea, but there are enough
bugs and different behaviors when doing so that it's clear `JavacTask`
was never meant to be used that way. Switching back to `call` with an
exception to abort early when just building an ABI. For pipelining
we'll do this call on another thread so that we can pause in the middle.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 1d62d4b