Decouple `DefaultClassUsageFileWriter` and `ClassUsageTracker`
Summary:
These two were coupled before because of the hacky way that dep file information moved through
the system when it was first implemented. That got un-hacked a while back, so this coupling is
no longer necessary, and it was getting in the way of outputting dep file information to
two locations in a pipelined build.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: c90151f