Revert D6097465: Move .buckd/tmp/ to buck-out/tmp/buckd/.
Summary:
This breaks users who want to be able to `rm -rf buck-out` while the daemon is
still running. This broke in two ways:

- `buck-out/tmp/buckd` is specified as `-Djava.io.tmpdir`. When the user runs
  Buck after deleting `buck-out`, `Files.createTempDirectory()` will fail
  because nothing recreates `buck-out/tmp/buckd/` when Buck contacts the daemon.
- When Buck is run as a pex, `com.facebook.buck.jvm.java.plugin.PluginLoader`
  unpacks `javac-plugin.jar` into a temp directory and stores that path in a
  static variable that persists for the lifetime of buckd. Because D6097465
  changed things so that file is under `buck-out`, after the user runs
  `rm -rf buck-out`, the daemon fails to load the plugin in the subsequent use
  of Buck.

I created D6111178 to address the first issue, but it did not fix the second.
Although I might be able to come up with a workaround for the second issue, it's
hard to safeguard against others creating a similar type of issue in the future.
Until I have a robust solution, the safest thing to do is to rollback.

Test Plan: revert-hammer

Reviewed By: styurin

fbshipit-source-id: e109430