Fix heisenbug with crashes in `finalizeBuildRule`
Summary:
When building a rule pipeline (even in the fake ABI pipelining that's in
the system today), `CachingBuildRuleBuilder.finalizeBuildRule` ends up running on a different thread
from the code that actually built the rule. If `finalizeBuildRule` throws,
`buildOrFetchFromCache` for the next rule in the pipeline will skip the
step where it waits for the pipelined build to complete and move on with
processing the canceled rule. Depending on how fast the pipelined build
completes, processing the canceled rule may call `BuildRuleScopeManager.scope()`
while the pipelined build has a scope open, resulting in an `IllegalStateException`.

The fix is to change `BuildRulePipelinesRunner.removeRule` (which is called no matter
what) to wait for the rule to complete.

Test Plan:
Introduced a failure in `finalizeBuildRule`, observed the failure was cleanly
displayed.

Reviewed By: dreiss

fbshipit-source-id: 58bb857