Refactor Dot.java
Summary:
Dot.java originally has two methods of exporting out the dot serialization of a Graph, a static and a non-static method along with a constructor.
We refactor Dot class to have only one non-static method to export the dot serialization, and change the constructor to require only the mandatory parameters.
Additional parameters can be set with a builder pattern

Test Plan: Unit Tests

Reviewed By: sbalabanov

fbshipit-source-id: 03e0c9a