Introduce source ABI pipelining
Summary:
This commit finally hooks up source ABI pipelining, which allows a single invocation
of `javac` to produce both a source ABI jar (after the frontend runs) and a
full jar (at the end). Since the frontend accounts for roughly half the time in the compiler,
this should cut bottlenecks due to large Java rules in half.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 6efe6fd