Detect missing grand-supers and grand-interfaces
Summary:
If `javac` fails to resolve a superclass or interface of a class being compiled, it's fine.
It will assume that the class/interface will be generated by an annotation processor.

If it *is* able to resolve a superclass or interface, however, it expects to be able
to fully resolve that superclass/interface, including all of its transitive superclasses/interfaces.
If it can't, you get class file not found errors.

This commit updates `InterfaceValidator` to detect such cases during full compilation so that
migration to source-only ABI can go more smoothly.

Test Plan: New unit test

Reviewed By: dreiss

fbshipit-source-id: 7c409e4