InterfaceScanner: Report only type references that affect the interface
Summary:
This fixes a couple ways in which `InterfaceScanner` was reporting type references incorrectly:
- type references in variable initializers are never part of the interface
- type references as part of compile-time constants (wherever they appear) are not part of the interface because the constant gets inlined
- a qualified reference to a member type is referencing the member type, but `InterfaceScanner` was reporting the topmost type instead

That last one was reported on the theory that we just needed to look for imported types, but it turns out
we care whether member types are referenced canonically or not, and that requires us to
look at the entire type reference. Subsequent commits will address that.

Test Plan:
Updated unit test for `InterfaceScanenr`
Run `InterfaceValidatorTest`

Reviewed By: dreiss

fbshipit-source-id: 781c182