InterfaceValidator: Type being available is not enough
Summary:
Source-only ABI gen needs to have canonical names of member types in order to generate
the `InnerClasses` table properly. I'm working towards having `InterfaceValidator` detect
non-canonical type references and recommend a fix.

This commit recognizes that just because the referenced type is available is not
enough for the refernce to be OK, but does not yet do anything about recommending
a reasonable solution.

Test Plan: Existing unit tests

Reviewed By: dreiss

fbshipit-source-id: 97d4239