InterfaceScanner: report static imports of canonical names as regular imports
Summary:
Static imports can be used to import a member type via a non-canonical name; in such cases we
consider the type to not be imported because we need the canonical name. However, when
a static import happens to use the canonical name we can consider the type imported.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: e2f2ac0