Minions now signal to coordinator as soon as a target is finished building, rather than in batches.
Summary:
The existing behavior is that minions fetch a batch of targets to build, build the entire batch, and then signal back to the coordinator that all targets in the batch are finished.

This is inefficient as it means that all dependent targets are blocked until the batch is finished, and also wastes resources (free CPU cores) at the minion whilst the batch is waiting to finish.

The new behavior is that minions will free up CPU cores as soon as a target has finished building, and will signal to the coordinator as soon as the target has been uploaded to cache.

Test Plan: ci

Reviewed By: shivanker, ruibm

fbshipit-source-id: e36c93a